refactor: split spline and filter utilities
