+++ main.js
@@ -3,6 +3,8 @@ import { initAdjustPanel, initLayerPanel, setAdjustCallbacks, setLayerCallbacks 
 import { updateStatus, updateZoom } from './gui/statusbar.js';
 import { Engine } from './engine.js';
 import { bmp, clipCanvas, layers, activeLayer, flattenLayers, renderLayers, updateLayerList, addLayer, deleteLayer } from './layer.js';
+import { catmullRomSpline } from './spline.js';
+import { applyFilterToCanvas } from './filters.js';
 
 window.bmp = bmp;
 
@@ -384,39 +386,6 @@ function drawEllipsePath(ctx, cx, cy, rx, ry) {
 
 
 
-function catmullRom(p0, p1, p2, p3, t) {
-  const t2 = t * t,
-    t3 = t2 * t;
-  return {
-    x:
-      0.5 *
-      (2 * p1.x +
-        (-p0.x + p2.x) * t +
-        (2 * p0.x - 5 * p1.x + 4 * p2.x - p3.x) * t2 +
-        (-p0.x + 3 * p1.x - 3 * p2.x + p3.x) * t3),
-    y:
-      0.5 *
-      (2 * p1.y +
-        (-p0.y + p2.y) * t +
-        (2 * p0.y - 5 * p1.y + 4 * p2.y - p3.y) * t2 +
-        (-p0.y + 3 * p1.y - 3 * p2.y + p3.y) * t3),
-  };
-}
-function catmullRomSpline(pts, seg = 16) {
-  const out = [];
-  if (pts.length < 2) return pts;
-  for (let i = 0; i < pts.length - 1; i++) {
-    const p0 = pts[i - 1] || pts[i],
-      p1 = pts[i],
-      p2 = pts[i + 1],
-      p3 = pts[i + 2] || p2;
-    for (let j = 0; j <= seg; j++) {
-      const t = j / seg;
-      out.push(catmullRom(p0, p1, p2, p3, t));
-    }
-  }
-  return out;
-}
 
 
 function makeCatmull(store) {
@@ -525,93 +494,8 @@ function makeCatmull(store) {
   };
 }
 
-function bspline(points, deg = 3, seg = 32) {
-  const n = points.length - 1;
-  if (n < deg) return points;
-  const knots = [];
-  const m = n + deg + 1;
-  for (let i = 0; i <= m; i++) {
-    if (i <= deg) knots.push(0);
-    else if (i >= n + 1) knots.push(n - deg + 1);
-    else knots.push(i - deg);
-  }
-  const out = [];
-  const start = knots[deg],
-    end = knots[n + 1];
-  const step = (end - start) / seg;
-  for (let s = 0; s <= seg; s++) {
-    const u = start + s * step;
-    let j = n;
-    if (u < end) {
-      for (let k = deg; k <= n; k++) {
-        if (u >= knots[k] && u < knots[k + 1]) {
-          j = k;
-          break;
-        }
-      }
-    }
-    out.push(deBoor(deg, u, knots, points, j));
-  }
-  return out;
-}
-function deBoor(k, u, t, c, j) {
-  const d = [];
-  for (let r = 0; r <= k; r++) d[r] = { ...c[j - k + r] };
-  for (let r = 1; r <= k; r++) {
-    for (let i = k; i >= r; i--) {
-      const idx = j - k + i;
-      const alpha = (u - t[idx]) / (t[idx + k + 1 - r] - t[idx]);
-      d[i] = {
-        x: (1 - alpha) * d[i - 1].x + alpha * d[i].x,
-        y: (1 - alpha) * d[i - 1].y + alpha * d[i].y,
-      };
-    }
-  }
-  return d[k];
-}
 
 
-function nurbs(points, weights, deg = 3, seg = 32) {
-  const n = points.length - 1;
-  // （そのままでも動くが、端点補間が欲しければクランプに置換する）
-  const knots = [];
-  for (let i = 0; i <= n + deg + 1; i++) knots.push(i);
-
-  function N(i, k, u) {
-    // ★ 末端uで最後の基底だけ1にする（w=0防止）
-    if (k === 0) {
-      if (u === knots[knots.length - 1]) return i === n ? 1 : 0;
-      return u >= knots[i] && u < knots[i + 1] ? 1 : 0;
-    }
-    const den1 = knots[i + k] - knots[i];
-    const den2 = knots[i + k + 1] - knots[i + 1];
-    const a = den1 ? (u - knots[i]) / den1 : 0;
-    const b = den2 ? (knots[i + k + 1] - u) / den2 : 0;
-    return a * N(i, k - 1, u) + b * N(i + 1, k - 1, u);
-  }
-
-  const uStart = knots[deg];
-  const uEnd = knots[n + 1];
-  const out = [];
-  // ★ 末端を含めない（代わりに最後だけ一歩手前でサンプル）
-  const step = (uEnd - uStart) / seg;
-  for (let s = 0; s <= seg; s++) {
-    // s==seg のときはごくわずか手前に寄せる
-    const u = s === seg ? uEnd - 1e-9 : uStart + s * step;
-    let x = 0,
-      y = 0,
-      w = 0;
-    for (let i = 0; i <= n; i++) {
-      const b = N(i, deg, u) * (weights[i] ?? 1);
-      x += points[i].x * b;
-      y += points[i].y * b;
-      w += b;
-    }
-    if (!isFinite(w) || Math.abs(w) < 1e-8) continue; // ★ 念のためNaN/ゼロガード
-    out.push({ x: x / w, y: y / w });
-  }
-  return out;
-}
 
 
 
@@ -1099,113 +983,6 @@ function applyFilter() {
     engine.requestRepaint();
   }
 }
-
-/* ===== Filter core (CPU) ===== */
-function applyFilterToCanvas(srcCanvas, p) {
-  const w = srcCanvas.width,
-    h = srcCanvas.height;
-  const out = document.createElement("canvas");
-  out.width = w;
-  out.height = h;
-  const sctx = srcCanvas.getContext("2d");
-  const dctx = out.getContext("2d");
-  const img = sctx.getImageData(0, 0, w, h);
-  const d = img.data;
-
-  const b = p.brightness / 100; // add in [ -1 .. 1 ]
-  //const c = Math.tan(((p.contrast / 100) * Math.PI) / 4); // contrast factor
-  const c = 1 + (p.contrast / 100); // 0 → 1.0（無変化）
-  const sat = 1 + p.saturation / 100; // multiply
-  const hue = ((p.hue || 0) * Math.PI) / 180; // radians
-  const inv = p.invert ? 1 : 0;
-
-
-  for (let i = 0; i < d.length; i += 4) {
-    let r = d[i] / 255,
-      g = d[i + 1] / 255,
-      bch = d[i + 2] / 255,
-      a = d[i + 3] / 255;
-
-    // brightness (add)
-    r = r + b;
-    g = g + b;
-    bch = bch + b;
-
-    // contrast around 0.5 -> remap to [-1..1] domain centered at 0
-    r = 0.5 + c * (r - 0.5);
-    g = 0.5 + c * (g - 0.5);
-    bch = 0.5 + c * (bch - 0.5);
-
-    // to HSV for hue/sat
-    let {
-      h: sH,
-      s: sS,
-      v: sV,
-    } = rgb2hsv(clamp01(r), clamp01(g), clamp01(bch));
-    sH = (sH + hue / (2 * Math.PI)) % 1;
-    if (sH < 0) sH += 1;
-    sS = clamp01(sS * sat);
-    ({ r, g, b: bch } = hsv2rgb(sH, sS, sV));
-
-    // invert
-    if (inv) {
-      r = 1 - r;
-      g = 1 - g;
-      bch = 1 - bch;
-    }
-
-    d[i] = Math.round(clamp01(r) * 255);
-    d[i + 1] = Math.round(clamp01(g) * 255);
-    d[i + 2] = Math.round(clamp01(bch) * 255);
-    d[i + 3] = Math.round(clamp01(a) * 255);
-  }
-  dctx.putImageData(img, 0, 0);
-  return out;
-}
-
-function clamp01(x) {
-  return x < 0 ? 0 : x > 1 ? 1 : x;
-}
-
-function rgb2hsv(r, g, b) {
-  const max = Math.max(r, g, b),
-    min = Math.min(r, g, b);
-  const d = max - min;
-  let h = 0;
-  if (d !== 0) {
-    if (max === r) h = ((g - b) / d) % 6;
-    else if (max === g) h = (b - r) / d + 2;
-    else h = (r - g) / d + 4;
-    h /= 6;
-    if (h < 0) h += 1;
-  }
-  const s = max === 0 ? 0 : d / max;
-  const v = max;
-  return { h, s, v };
-}
-
-function hsv2rgb(h, s, v) {
-  const i = Math.floor(h * 6),
-    f = h * 6 - i,
-    p = v * (1 - s),
-    q = v * (1 - f * s),
-    t = v * (1 - (1 - f) * s);
-  switch (i % 6) {
-    case 0:
-      return { r: v, g: t, b: p };
-    case 1:
-      return { r: q, g: v, b: p };
-    case 2:
-      return { r: p, g: v, b: t };
-    case 3:
-      return { r: p, g: q, b: v };
-    case 4:
-      return { r: t, g: p, b: v };
-    case 5:
-      return { r: v, g: p, b: q };
-  }
-}
-
 
 //#endregion
